Integrate classic comm log with new log
