+++ frontend/src/components/CommunicationLog/constants.js
@@ -11,6 +11,7 @@ const defaultLogValues = {
   duration: '',
   method: '',
   purpose: '',
+  result: '',
 };
 
 const defaultAttachmentValues = {
@@ -54,6 +55,33 @@ const pageComplete = (
   fields,
 ) => fields.every((field) => hookForm.getValues(field));
 
+/**
+   * this is just a simple handler to "flatten"
+   * the JSON column data into the form
+   *
+   * @param {fn} reset this is the hookForm.reset function (pass it a new set of values and it
+   *  replaces the form with those values; it also calls the standard form.reset event
+   * @param {*} updatedLog - the log object from the database, which has some
+   * information stored at the top level of the object, and some stored in a data column
+   */
+const resetFormData = (reset, updatedLog) => {
+  const {
+    data,
+    updatedAt,
+    recipients,
+    ...fields
+  } = updatedLog;
+
+  const form = {
+    ...defaultValues,
+    ...data,
+    ...fields,
+    recipients: recipients.map((r) => ({ value: r.id, label: r.name })),
+  };
+
+  reset(form);
+};
+
 export {
   defaultValues,
   defaultLogValues,
@@ -62,4 +90,5 @@ export {
   formatRegionalCommunicationLogUrl,
   pageComplete,
   nextStepsFields,
+  resetFormData,
 };
